+++ src/main/java/ch/qligier/jetbrains/plugin/fss/igini/IgIniSpecs.java
@@ -1,0 +1,21 @@
+// Copyright 2024 Quentin Ligier. Use of this source code is governed by the MIT license.
+
+package ch.qligier.jetbrains.plugin.fss.igini;
+
+/**
+ * Specifications of the ig.ini file.
+ *
+ * @author Quentin Ligier
+ **/
+public class IgIniSpecs {
+
+    public static final String IG_SECTION_NAME = "[IG]";
+    public static final String IG_KEY_NAME = "ig";
+    public static final String TEMPLATE_KEY_NAME = "template";
+
+    /**
+     * This class is not instantiable.
+     */
+    private IgIniSpecs() {
+    }
+}