+++ src/main/java/ch/qligier/jetbrains/plugin/fss/igini/language/parser/IniParserUtil.java
@@ -1,0 +1,13 @@
+// Copyright 2024 Quentin Ligier. Use of this source code is governed by the MIT license.
+
+package ch.qligier.jetbrains.plugin.fss.igini.language.parser;
+
+import com.intellij.lang.parser.GeneratedParserUtilBase;
+
+/**
+ * Some utilities used by the generated parser.
+ *
+ * @author Quentin Ligier
+ **/
+public class IniParserUtil extends GeneratedParserUtilBase {
+}