+++ src/main/java/ch/qligier/jetbrains/plugin/fss/igini/language/psi/IniValue.java
@@ -1,0 +1,10 @@
+// This is a generated file. Not intended for manual editing.
+package ch.qligier.jetbrains.plugin.fss.igini.language.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public interface IniValue extends PsiElement {
+
+}